Moving `ColorConverter` and `MaterialHelper` classes
