+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/game/Team.java
@@ -25,7 +25,7 @@ import de.butzlabben.missilewars.game.enums.TeamType;
 import de.butzlabben.missilewars.game.misc.TeamSpawnProtection;
 import de.butzlabben.missilewars.menus.MenuItem;
 import de.butzlabben.missilewars.player.MWPlayer;
-import de.butzlabben.missilewars.util.version.ColorConverter;
+import de.butzlabben.missilewars.util.ColorConverter;
 import de.redstoneworld.redutilities.player.Teleport;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
