+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/listener/game/GameListener.java
@@ -35,7 +35,7 @@ import de.butzlabben.missilewars.game.schematics.objects.Missile;
 import de.butzlabben.missilewars.listener.ShieldListener;
 import de.butzlabben.missilewars.menus.inventory.TeamSelectionMenu;
 import de.butzlabben.missilewars.player.MWPlayer;
-import de.butzlabben.missilewars.util.version.MaterialHelper;
+import de.butzlabben.missilewars.util.MaterialHelper;
 import org.bukkit.*;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
